+++ kleidicv_thread/include/kleidicv_thread/kleidicv_thread.h
@@ -265,6 +265,16 @@ kleidicv_error_t kleidicv_thread_saturating_add_abs_with_threshold_s16(
 
 /// Internal - not part of the public API and its direct use is not supported.
 ///
+/// Multithreaded implementation of kleidicv_rotate - see the documentation of
+/// that function for more details.
+kleidicv_error_t kleidicv_thread_rotate(const void *src, size_t src_stride,
+                                        size_t width, size_t height, void *dst,
+                                        size_t dst_stride, int angle,
+                                        size_t element_size,
+                                        kleidicv_thread_multithreading);
+
+/// Internal - not part of the public API and its direct use is not supported.
+///
 /// Multithreaded implementation of kleidicv_gaussian_blur_u8 - see the
 /// documentation of that function for more details.
 kleidicv_error_t kleidicv_thread_gaussian_blur_u8(